move CountrySettings UI to settings/country package
