+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.country_CountrySettingsLayoutPaparazziTest_CountrySettingsLayout()[Nexus_5,NOTNIGHT,ACCESSIBILITY].png
