+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.country_CountrySettingsLayoutPaparazziTest_CountrySettingsLayout()[Pixel_6_Pro,NIGHT,NO_ACCESSIBILITY].png
