+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.country_CountrySettingsLayoutPaparazziTest_CountrySettingsLayout()_-_ConfirmationOverlay[Nexus_5,NOTNIGHT,NO_ACCESSIBILITY].png
