+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.country_CountrySettingsLayoutPaparazziTest_CountrySettingsLayout()_-_Searching[Nexus_5,NOTNIGHT,NO_ACCESSIBILITY].png
