+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.country_CountrySettingsLayoutPaparazziTest_CountrySettingsLayout()_-_Selected_Country[Nexus_5,NOTNIGHT,NO_ACCESSIBILITY].png
