+++ app/src/test/snapshots/images/org.cru.godtools.ui.settings.country_CountrySettingsLayoutPaparazziTest_CountrySettingsLayout()_-_Selected_Country[Pixel_6_Pro,NOTNIGHT,NO_ACCESSIBILITY].png
